requirements.txt
clickhouse-driver==0.2.4
mysql-connector-python==8.0.31
sqlparse>=0.4.3